Update some examples to not use X and import from shiny.ui
